Leverage SCC for Class Relationship Verification
Record and process class relationship snippets, which indicate the
indices of the source and target classes of a relationship in the
classNameList of verifyData.

Store class relationship snippets for a romClass into the shared
class cache during the first run of an application. Snippets are
stored for classes that have passed bytecode verification. On
subsequent runs of the application, snippets will already exist in
the cache for the romClass, which means that the class's bytecodes
had already been verified and thus we can skip the linear bytecode
walk.

This work is an extension of the existing
-XX:+ClassRelationshipVerifier work. It is enabled/disabled using
the same command-line options.

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>